Adjust workflows for publishing multi-arch images
The build stack process does not need to change. The same command will now just produce a multi-arch image.

Unfortunately, skopeo cannot copy a multi-arch image, so we use a custom `jam publish-stack` command to do this.

This commit adds a wrapper `publish.sh` script which publishes the images. The wrapper script is convenient for running locally. It also keeps the inline script smaller in the workflow file. This script takes in a list of build refs, a list of run refs and the build and run archive files. It then iterates over the pairs of run-ref & build-ref (in order) and publishes them. This makes it easy to specify multiple registries/tags and publish them all in one go.

Signed-off-by: Daniel Mikusa <dan@mikusa.com>