Fix tests failing because stack is now multi-arch
- Modify the metadata related tests to confirm both architectures
- Modify integration tests so that we don't use skopeo, which can't handle multi-arch images

The latter was a little tricky. I couldn't find a way to load the multi-arch OCI TAR file into the Docker daemon so we could run the test. Instead, I had to spin up a Registry running in Docker, then load the multi-arch OCI TAR files for the stack into the internal Registry. Then we can build from the registry.

Signed-off-by: Daniel Mikusa <dan@mikusa.com>