Arm64 (#108)
* Add ARM64 stack

Signed-off-by: Daniel Mikusa <dan@mikusa.com>

* Adjust workflows for publishing multi-arch images

The build stack process does not need to change. The same command will now just produce a multi-arch image.

Unfortunately, skopeo cannot copy a multi-arch image, so we use a custom `jam publish-stack` command to do this.

This commit adds a wrapper `publish.sh` script which publishes the images. The wrapper script is convenient for running locally. It also keeps the inline script smaller in the workflow file. This script takes in a list of build refs, a list of run refs and the build and run archive files. It then iterates over the pairs of run-ref & build-ref (in order) and publishes them. This makes it easy to specify multiple registries/tags and publish them all in one go.

Signed-off-by: Daniel Mikusa <dan@mikusa.com>

* Set up QEMU

Set up QEMU for building multi-architectures. This was explored https://github.com/paketo-buildpacks/jammy-tiny-stack/pull/111/commits/31c6477d61f974b749fee761d570376a376ee539.

Signed-off-by: Daniel Mikusa <dan@mikusa.com>

* Fix tests failing because stack is now multi-arch

- Modify the metadata related tests to confirm both architectures
- Modify integration tests so that we don't use skopeo, which can't handle multi-arch images

The latter was a little tricky. I couldn't find a way to load the multi-arch OCI TAR file into the Docker daemon so we could run the test. Instead, I had to spin up a Registry running in Docker, then load the multi-arch OCI TAR files for the stack into the internal Registry. Then we can build from the registry.

Signed-off-by: Daniel Mikusa <dan@mikusa.com>

---------

Signed-off-by: Daniel Mikusa <dan@mikusa.com>